+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [0.7.0] - Unreleased
--
+- Added option to not show shell commands in the interactive view.
 
 ## [0.6.0] - 2021-02-09
 - Added bold label display option.